updated_requirements.txt
pyyaml
numpy
pandas
scikit-learn
pyarrow
flake8
pytest
black
mypy
nox
mlflow
matplotlib
prefect
torch
torchvision
pytorch-lightning
fastapi
uvicorn

